pyproject.toml
[tool.poetry]
name = "repl_python3_beta_blackjack-start"
version = "0.1.0"
description = ""
authors = ["hanane84 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
repl = "^1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
